# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "otelo",
    "description": "Query the logs, traces, and metrics that otelo keeps. Every path but the login and the logout needs the password of `otelo init`, as the cookie `otelo_password` or as `Authorization: Bearer <password>`.",
    "license": {
      "name": ""
    },
    "version": "0.0.1"
  },
  "paths": {
    "/api/attributes": {
      "get": {
        "summary": "The attribute keys of the records of a signal and of their resources over\nthe retention, the most common first, with their types.",
        "operationId": "list_attribute_keys",
        "parameters": [
          {
            "name": "signal",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Signal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributeKeys"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/complete": {
      "get": {
        "summary": "Suggests the fields, operators, values, and keywords that can go at the\ncursor of a query, from the attributes and values of the retention, and\ndescribes the field of the term the cursor is in.",
        "operationId": "complete_query",
        "parameters": [
          {
            "name": "signal",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Signal"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The query as typed so far.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "The position of the cursor in the query, in characters. The end of the\nquery when missing.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Completions"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/indexes/{signal}/{key}": {
      "put": {
        "summary": "Indexes an attribute of the logs or of the spans, so a query that\ncompares it reads only the matching records. The indexer builds\nthe index within seconds.",
        "operationId": "add_index",
        "parameters": [
          {
            "name": "signal",
            "in": "path",
            "description": "`logs` or `spans`",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IndexedSignal"
            }
          },
          {
            "name": "key",
            "in": "path",
            "description": "The attribute key, such as `user.id`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexList"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Drops the index of an attribute.",
        "operationId": "remove_index",
        "parameters": [
          {
            "name": "signal",
            "in": "path",
            "description": "`logs` or `spans`",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IndexedSignal"
            }
          },
          {
            "name": "key",
            "in": "path",
            "description": "The attribute key, such as `user.id`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexList"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/logs": {
      "get": {
        "summary": "Log lines, newest first.",
        "operationId": "list_logs",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most rows to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The records to keep, such as `http.route = \"/matches\" OR user.id = 7`.\nEvery record when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Logs"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/groups": {
      "get": {
        "summary": "Log lines grouped by message template, the largest group first. The\ntemplate replaces numbers, UUIDs, hex IDs, and quoted strings with\nplaceholders.",
        "operationId": "list_log_groups",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most rows to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The records to keep, such as `http.route = \"/matches\" OR user.id = 7`.\nEvery record when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogGroups"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/services": {
      "get": {
        "summary": "The services that sent spans or logs in the range, the most requests\nfirst. A request is a span of the server kind with\n`http.request.method`. Each service has its requests and its logs over\nthe range and in buckets of one step, and the count of its spans.",
        "operationId": "list_services",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most services to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "step",
            "in": "query",
            "description": "The length of a bucket, such as `1m`. One that makes 60 buckets at\nmost when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Services"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{name}": {
      "get": {
        "summary": "One service: its requests and logs over the range and in buckets of one\nstep. A service without telemetry in the range has none of either.",
        "operationId": "get_service",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The name of the service",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "step",
            "in": "query",
            "description": "The length of a bucket, such as `1m`. One that makes 120 buckets at\nmost when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/spans": {
      "get": {
        "summary": "Spans, newest first.",
        "operationId": "list_spans",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most rows to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The records to keep, such as `http.route = \"/matches\" OR user.id = 7`.\nEvery record when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spans"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/spans/groups": {
      "get": {
        "summary": "The spans that the query keeps, grouped by the values of the `by` names,\nthe most time first. Each group has the count, the failures, the total\ntime, and the latency percentiles of its spans, and the name and the\nattributes of its newest span. The answer has the same numbers for all the\nspans, over the range and in buckets of one step.",
        "operationId": "list_span_groups",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most groups to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The spans to keep, such as `service = \"api\" kind = server`. Every span\nwhen missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "by",
            "in": "query",
            "description": "The names to group the spans by, separated by commas: attributes,\n`service`, `name`, or `resource.<key>`, such as\n`http.request.method,http.route`. All spans are one group when\nmissing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "step",
            "in": "query",
            "description": "The length of a bucket, such as `1m`. One that makes 120 buckets at\nmost when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpanGroups"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces": {
      "get": {
        "summary": "Traces with a span that the query keeps, by their root span, newest first.",
        "operationId": "list_traces",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. One hour before `until` when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most rows to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "The records to keep, such as `http.route = \"/matches\" OR user.id = 7`.\nEvery record when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Traces"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{trace_id}": {
      "get": {
        "summary": "One trace: its spans by start time, and the logs that carry its ID.",
        "operationId": "get_trace",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "description": "The trace ID as 32 hex digits",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "The start of the range: a duration before now, such as `1h`, or an\nRFC 3339 timestamp. The whole retention when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "The end of the range, in the form of `since`. Now when missing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most spans, and the most logs, to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trace"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AttributeKeys": {
        "type": "object",
        "description": "The attribute keys of a signal, over the retention of the signal.",
        "required": [
          "record",
          "resource"
        ],
        "properties": {
          "record": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attribute"
            },
            "description": "The attributes of the records: of the logs, the spans, or the attributes\nof the series."
          },
          "resource": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attribute"
            },
            "description": "The attributes of the resources that sent them."
          }
        }
      },
...
      "Completions": {
        "type": "object",
        "description": "What can go at the cursor of a query, and what the field there holds.",
        "required": [
          "suggestions",
          "field"
        ],
        "properties": {
          "field": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FieldBody",
                "description": "The field of the term the cursor is in. Missing when the cursor is in\nno term, and for an attribute that no record has."
              }
            ]
          },
          "suggestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuggestionBody"
            }
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "IndexBody": {
        "type": "object",
        "required": [
          "signal",
          "key"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "signal": {
            "$ref": "#/components/schemas/IndexedSignal",
            "description": "`logs` or `spans`."
          }
        }
      },
      "IndexList": {
        "type": "object",
        "description": "The attributes that have an index.",
        "required": [
          "indexes"
        ],
        "properties": {
          "indexes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexBody"
            }
          }
        }
      },
      "IndexedSignal": {
        "type": "string",
        "description": "The kind of record whose attributes can have an index, as the API names\nit.",
        "enum": [
          "logs",
          "spans"
        ]
      },
...
      "Logs": {
        "type": "object",
        "description": "Log lines, newest first.",
        "required": [
          "logs",
          "truncated",
          "unindexed"
        ],
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogLine"
            }
          },
          "truncated": {
            "type": "boolean",
            "description": "More lines match than the limit let through."
          },
          "unindexed": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The attributes the query compares that have no index, so it read every\nline in the range."
          }
        }
      },
...
      "Signal": {
        "type": "string",
        "description": "The kind of record a query reads, as the API names it.",
        "enum": [
          "logs",
          "spans",
          "metrics"
        ]
      },
...
    }
  }
}